refactor: ExpertCategory 타입 정의 및 import 경로 수정 (#94)
- ExpertCategory 타입을 '@/types/expert'로 이동하여 코드의 모듈화 및 재사용성 향상
- 관련 컴포넌트에서 import 경로를 업데이트하여 일관성 유지
